+++ melody/helpers/_dataclass.py
@@ -18,6 +18,7 @@ class Media:
     time: int = 0
     user: str = None
     video: bool = False
+    is_live: bool = False
 
 
 @dataclass