+++ melody/plugins/active.py
@@ -26,7 +26,8 @@ async def _activevc(_, m: types.Message):
 
     for i, chat in enumerate(db.active_calls):
         playing = queue.get_current(chat)
-        text += f"\n{i+1}. <code>{chat}</code>\n    ➜ {playing.title[:25]}"
+        title = playing.title[:25] if playing and playing.title else "—"
+        text += f"\n{i+1}. <code>{chat}</code>\n    ➜ {title}"
 
     if len(text) < 4000:
         return await sent.edit_text(m.lang["vc_list"] + text)